requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
psycopg2-binary==2.9.9
tensorflow==2.14.0
Pillow==10.0.1
python-dotenv==1.0.0
requests==2.31.0
folium==0.14.0
geopy==2.4.0
opencv-python==4.8.1.78
gunicorn==21.2.0
SQLAlchemy==2.0.23
alembic==1.12.1
